+++ src/utils/util.ts
@@ -331,31 +331,6 @@ export function boxInBox(
   else return boundH * childAR;
 }
 
-export function getCallStack(): string[] {
-  return new Error()
-    .stack!.split("\n")
-    .map((ln) => ln.trim())
-    .filter((ln) => ln !== "Error" && !ln.includes("getCallStack"));
-}
-
-let blameMaps = new Map<string, Map<string, number>>();
-export function dbgAddBlame(kind: string, amount: number) {
-  let map = blameMaps.get(kind);
-  if (!map) {
-    map = new Map<string, number>();
-    blameMaps.set(kind, map);
-  }
-  getCallStack().forEach((ln) => {
-    map!.set(ln, (map!.get(ln) ?? 0) + amount);
-  });
-}
-export function dbgGetBlame(kind: string) {
-  return blameMaps.get(kind);
-}
-export function dbgClearBlame(kind: string) {
-  blameMaps.get(kind)?.clear();
-}
-
 // TODO(@darzu): would be nice to have a dbg mode where we assert that this isn't
 //    being triggered at different lengths at a callsite in the steady state
 // TODO(@darzu): actually we probably don't want this fn, better to just do a while loop + assert length hasn't shrunk?
